bulk sync
